refactor: uniform cog loader
